Move test doubles to its own namespace
