+++ packages/contentsgarten/src/index.ts
@@ -5,3 +5,4 @@ export * from './ContentsgartenCache'
 export * from './ContentsgartenRouter'
 export * from './ContentsgartenTeamResolver'
 export * from './GitHubApp'
+export * from './testing'
